+++ modules/core/src/dxt.cpp
@@ -2164,7 +2164,7 @@ static bool ocl_mulSpectrums( InputArray _srcA, InputArray _srcB,
     k.args(ocl::KernelArg::ReadOnlyNoSize(A), ocl::KernelArg::ReadOnlyNoSize(B),
            ocl::KernelArg::WriteOnly(dst));
 
-    size_t globalsize[2] = { asize.width, asize.height };
+    size_t globalsize[2] = { static_cast<size_t>(asize.width), static_cast<size_t>(asize.height) };
     return k.run(2, globalsize, NULL, false);
 }
