+++ modules/core/src/matrix.cpp
@@ -2701,7 +2701,7 @@ static bool ocl_setIdentity( InputOutputArray _m, const Scalar& s )
     UMat m = _m.getUMat();
     k.args(ocl::KernelArg::WriteOnly(m), ocl::KernelArg::Constant(Mat(1, 1, type, s)));
 
-    size_t globalsize[2] = { m.cols, m.rows };
+    size_t globalsize[2] = { static_cast<size_t>(m.cols), static_cast<size_t>(m.rows) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -2934,7 +2934,7 @@ static bool ocl_transpose( InputArray _src, OutputArray _dst )
                ocl::KernelArg::WriteOnlyNoSize(dst));
 
     size_t localsize[3]  = { TILE_DIM, BLOCK_ROWS, 1 };
-    size_t globalsize[3] = { src.cols, inplace ? src.rows : divUp(src.rows, TILE_DIM) * BLOCK_ROWS, 1 };
+    size_t globalsize[3] = { static_cast<size_t>(src.cols), static_cast<size_t>(inplace ? src.rows : divUp(src.rows, TILE_DIM) * BLOCK_ROWS), 1 };
 
     return k.run(2, globalsize, localsize, false);
 }
